move data to new data folder
