+++ src/data/trackData.tsx
@@ -1,0 +1,35 @@
+export type TrackData = {
+  title: string;
+  id: string;
+};
+
+export const tracks2020: TrackData[] = [
+  { title: "The New Rock and Roll", id: "1848517437" },
+  { title: "Likey Likey", id: "2587741727" },
+  { title: "Rise Up", id: "2036553205" },
+  { title: "That's Tasty", id: "171503560" },
+];
+
+export const tracks2019: TrackData[] = [
+  { title: "Kong Drop", id: "3545714909" },
+  { title: "WTF", id: "2726046130" },
+];
+
+export const tracks2018: TrackData[] = [
+  { title: "Tell Me Daddy", id: "2520571840" },
+  { title: "All I Wanna Do", id: "1563626637" },
+  { title: "Work It Move It Make It", id: "3855208459" },
+  { title: "Leroy Jenkins (Jump & Twist)", id: "359552405" },
+  { title: "Mangoes Bananas", id: "447342307" },
+];
+
+export const tracks2016: TrackData[] = [
+  { title: "Never Ever Land", id: "3974639248" },
+  { title: "Ju Ju Magik", id: "3974639248" },
+];
+
+export const tracks2015: TrackData[] = [
+  { title: "Club Tropigwana", id: "743691718" },
+  { title: "Remember Me", id: "3008176448" },
+  { title: "Get Up", id: "4132457639" },
+];
